+++ src/pull-down-refresh/type.ts
@@ -8,12 +8,20 @@ import { LoadingProps } from '../loading/index';
 
 export interface TdPullDownRefreshProps {
   /**
-   * 自定义组件样式
-   * @default ''
+   * iOS点击顶部状态栏、安卓双击标题栏时，滚动条返回顶部，只支持竖向。自 2.27.3 版本开始，若非显式设置为 false，则在显示尺寸大于屏幕 90% 时自动开启
+   * @default true
    */
-  style?: {
-    type: StringConstructor;
-    value?: string;
+  enableBackToTop?: {
+    type: BooleanConstructor;
+    value?: boolean;
+  };
+  /**
+   * 开启 passive 特性，能优化一定的滚动性能
+   * @default false
+   */
+  enablePassive?: {
+    type: BooleanConstructor;
+    value?: boolean;
   };
   /**
    * 加载loading样式
@@ -27,8 +35,7 @@ export interface TdPullDownRefreshProps {
    * @default 50
    */
   loadingBarHeight?: {
-    type: StringConstructor;
-    optionalTypes: Array<NumberConstructor>;
+    type: null;
     value?: string | number;
   };
   /**
@@ -47,12 +54,19 @@ export interface TdPullDownRefreshProps {
     value?: string[];
   };
   /**
+   * 距底部/右边多远时，触发 scrolltolower 事件
+   * @default 50
+   */
+  lowerThreshold?: {
+    type: null;
+    value?: string | number;
+  };
+  /**
    * 最大下拉高度，如果值为数字则单位是：'px'
    * @default 80
    */
   maxBarHeight?: {
-    type: StringConstructor;
-    optionalTypes: Array<NumberConstructor>;
+    type: null;
     value?: string | number;
   };
   /**
@@ -62,6 +76,22 @@ export interface TdPullDownRefreshProps {
   refreshTimeout?: {
     type: NumberConstructor;
     value?: number;
+  };
+  /**
+   * 值应为某子元素id（id不能以数字开头）。设置哪个方向可滚动，则在哪个方向滚动到该元素
+   * @default ''
+   */
+  scrollIntoView?: {
+    type: StringConstructor;
+    value?: string;
+  };
+  /**
+   * 距顶部/左边多远时，触发 scrolltoupper 事件
+   * @default 50
+   */
+  upperThreshold?: {
+    type: null;
+    value?: string | number;
   };
   /**
    * 组件状态，值为 `true` 表示下拉状态，值为 `false` 表示收起状态