+++ src/tabs/tabs.ts
@@ -63,8 +63,6 @@ export default class Tabs extends SuperComponent {
     tabs: [],
     currentIndex: -1,
     trackStyle: '',
-    isScrollX: true,
-    direction: 'X',
     offset: 0,
     tabID: '',
     placement: 'top',
@@ -95,6 +93,12 @@ export default class Tabs extends SuperComponent {
   }
 
   methods = {
+    onScroll(e) {
+      const { scrollLeft } = e.detail;
+      this.setData({
+        offset: scrollLeft,
+      });
+    },
     updateTabs(cb) {
       const { children } = this;
       const tabs = children.map((child: any) => child.data);
@@ -166,7 +170,7 @@ export default class Tabs extends SuperComponent {
       if (!this.properties.showBottomLine) return;
       const { children } = this;
       if (!children) return;
-      const { currentIndex, isScrollX, direction } = this.data;
+      const { currentIndex } = this.data;
       if (currentIndex <= -1) return;
 
       try {
@@ -179,10 +183,10 @@ export default class Tabs extends SuperComponent {
 
         res.forEach((item) => {
           if (count < currentIndex) {
-            distance += isScrollX ? item.width : item.height;
+            distance += item.width;
             count += 1;
           }
-          totalSize += isScrollX ? item.width : item.height;
+          totalSize += item.width;
         });
 
         if (this.containerWidth) {
@@ -193,18 +197,14 @@ export default class Tabs extends SuperComponent {
           });
         }
 
-        if (isScrollX && this.data.theme === 'line') {
+        if (this.data.theme === 'line') {
           const trackLineWidth = await this.getTrackSize();
           distance += (rect.width - trackLineWidth) / 2;
         }
-        let trackStyle = `-webkit-transform: translate${direction}(${distance}px);
-          transform: translate${direction}(${distance}px);
-        `;
-        if (!isScrollX) {
-          trackStyle += `height: ${rect.height}px;`;
-        }
         this.setData({
-          trackStyle,
+          trackStyle: `-webkit-transform: translateX(${distance}px);
+            transform: translateX(${distance}px);
+          `,
         });
       } catch (err) {
         this.triggerEvent('error', err);